fix: address P2 review finding — extract shared validation utilities
Moved COMMUNITIES_LOWER, LANGUAGE_CODES, VALID_LANGUAGES, and
validateCommunity into server/utils/validation.ts as the single source
of truth. metrics.ts and report.ts now import from the shared module.
generate-reports.ts extends the shared LANGUAGE_CODES with 'Other'.
This ensures consistent validation rigor (SQL wildcard stripping,
length limits, allowlist checks) across all routes.
